♻️ refactor: 레시피 메타데이터 수정
