+++ src/app/recipe/[_id]/edit/RecipeEditForm.tsx
@@ -7,7 +7,7 @@ import Button from '@/components/common/Button';
 import dynamic from 'next/dynamic';
 import 'react-quill-new/dist/quill.snow.css';
 import Swal from 'sweetalert2';
-import RecipeEditLoading from './EditLoading';
+import RecipeEditLoading from './Loading';
 import { updateRecipe } from '@/data/actions/recipe';
 
 const QuillNoSSRWrapper = dynamic(() => import('react-quill-new'), {
